設定刪除資料條件

+++ routes/address-book.js
@@ -179,8 +179,14 @@ router.delete("/api/:sid", async (req, res) => {
     code: 0,
     result: {}
   };
+
+  if(!req.my_jwt?.id){
+    output.code=470;
+    return res.json(output);
+  }
   const sid = +req.params.sid || 0;
   if (!sid) {
+    output.code=480;
     return res.json(output);
   }
   const sql = `DELETE FROM address_book WHERE sid=${sid}`;
